Add files via upload
A multi-purpose calculator. It contains a rudimentary calculator that will only add the two inputs, but also contains a BMI-calculator that will calculate ones BMI, based on the height and weight entered by the user.  

The project has to run locally on the device or on a server, in order to be operational. 

The following npm-modules are required and used in this project:
- express
- bodyParser